requirements.txt
# https://github.com/5j9/wikitextparser#templates
wikitextparser
# https://pypi.org/project/Jinja2/
jinja2
# https://pypi.org/project/pylodstorage/
pylodstorage